fix(ci): remove caching of node_modules for Travis
Yes, this makes Travis builds slower, but there's an issue in that the
npm postinstall script is not executed because there's no install
happening, as Travis recognizes that it doesn't need to run yarn.

Also, every time we change a dependency, we need to manually delete the
Travis cache if we don't want the build to fail.